+++ src/bin/pica/util.rs
@@ -7,7 +7,6 @@ pub(crate) type CliResult<T> = Result<T, CliError>;
 pub(crate) enum CliError {
     Io(io::Error),
     Csv(csv::Error),
-    Xml(xml::writer::Error),
     Pica(pica::Error),
     ParsePica(pica_record::ParsePicaError),
     ParsePath(pica_path::ParsePathError),
@@ -20,7 +19,6 @@ impl fmt::Display for CliError {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match *self {
             CliError::Csv(ref e) => e.fmt(f),
-            CliError::Xml(ref e) => e.fmt(f),
             CliError::Io(ref e) => e.fmt(f),
             CliError::Pica(ref e) => e.fmt(f),
             CliError::ParsePica(ref e) => e.fmt(f),
@@ -47,12 +45,6 @@ impl From<io::Error> for CliError {
 impl From<csv::Error> for CliError {
     fn from(err: csv::Error) -> CliError {
         CliError::Csv(err)
-    }
-}
-
-impl From<xml::writer::Error> for CliError {
-    fn from(err: xml::writer::Error) -> CliError {
-        CliError::Xml(err)
     }
 }