+++ tests/pica/mod.rs
@@ -5,7 +5,6 @@ use crate::TestResult;
 
 mod json;
 mod sample;
-mod xml;
 
 #[test]
 fn pica_io_error() -> TestResult {