+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* #637 Stabilize `print` command
+
 ## [0.17.0] - 2023-06-30
 
 ### Added