+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * #637 Stabilize `print` command
 * #641 Stabilize `sample` command
 * #642 Add `--squash` and `--merge` option
+* #644 Add `!^` and `!$` operator
 
 ### Changed