+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * #643 Print more helpful error message on `ParsePicaError` 
+* #653 Don't require filter argument when an expression file is given
 
 ### Removed