+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * #687 Allow multi-field queries in `frequency` command
 
 
+### Fixed
+
+* #688 Allow "not"-operands in or-expressions
+
+
 ## [0.18.0] - 2023-07-27
 
 ### Added